demonstration_generator/requirements.txt
numpy
h5py
opencv-python
Pillow
tqdm
torch